requirements.txt
pydantic>=2.0,<3.0
python-dateutil>=2.8
boto3>=1.28
pyarrow>=14.0
